Refactor imports to use @workers/versions for consistency
