+++ suites/functional/permissions.test.ts
@@ -1,6 +1,6 @@
 import { setupTestLifecycle } from "@helpers/vitest";
 import { getWorkers } from "@workers/manager";
-import { type Group } from "@xmtp/node-sdk";
+import { type Group } from "@workers/versions";
 import { describe, expect, it } from "vitest";
 
 const testName = "permissions";
