+++ suites/networkchaos/node-blackhole.test.ts
@@ -2,7 +2,7 @@ import { verifyMessageStream } from "@helpers/streams";
 import { setupTestLifecycle } from "@helpers/vitest";
 import { typeofStream } from "@workers/main";
 import { getWorkers } from "@workers/manager";
-import type { Group } from "@xmtp/node-sdk";
+import { type Group } from "@workers/versions";
 import { describe, expect, it } from "vitest";
 import { DockerContainer } from "../../network-stability-utilities/container";
